Modifications to the user registration endpoint and database in the cloud
We added in the input json of the enpoint data that the alpaca api does not contemplate, such as the password and the associated commission agent id. For this we created a request class that receives all the data required by alpaca to create an account and the additional data we need.

In addition, a MySQL instance was created in the cloud and then connected to the Backend.